rename test -> real-test add simple test adapt CI settings
