+++ packages/gist_backend/README.md
@@ -3,7 +3,8 @@
 The backend to submit gists to GitHub while keeping authorization private.
 
 Before starting this backend please set the TD_PLAYGROUND_LOGIN and TD_PLAYGROUND_TOKEN environment variables on your system to your GitHub access credentials.
+For testing purposes (to see if npm test succeeds) setting empty Variables is sufficient, to pass the `npm run realTest` command your credentials have to be valid.
 
-## License 
+## License
 
 Licensed under the MIT license, see [License](./LICENSE.md).
